# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_trackman_csv_api_v1_imports_trackman_csv_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_trackman_csv_api_v1_imports_trackman_csv_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportDetailOut": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "import_metadata": {
            "additionalProperties": true,
            "title": "Import Metadata",
            "type": "object"
          },
          "import_type": {
            "title": "Import Type",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ImportIssueOut"
            },
            "title": "Issues",
            "type": "array"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "rows_accepted": {
            "title": "Rows Accepted",
            "type": "integer"
          },
          "rows_rejected": {
            "title": "Rows Rejected",
            "type": "integer"
          },
          "rows_total": {
            "title": "Rows Total",
            "type": "integer"
          },
          "source_status": {
            "$ref": "#/components/schemas/SourceStatus"
          },
          "status": {
            "$ref": "#/components/schemas/ImportStatus"
          }
        },
        "required": [
          "id",
          "provider",
          "import_type",
          "filename",
          "status",
          "source_status",
          "rows_total",
          "rows_accepted",
          "rows_rejected",
          "created_at",
          "completed_at",
          "error_message"
        ],
        "title": "ImportDetailOut",
        "type": "object"
      },
      "ImportIssueCode": {
        "description": "Machine-readable reasons a row or file was rejected or flagged.",
        "enum": [
          "SCHEMA_UNKNOWN",
          "MISSING_REQUIRED_FIELD",
          "INVALID_VALUE",
          "INVALID_UNIT",
          "PLAYER_UNRESOLVED",
          "AMBIGUOUS_SESSION"
        ],
        "title": "ImportIssueCode",
        "type": "string"
      },
      "ImportIssueOut": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ImportIssueCode"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "row_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Number"
          }
        },
        "required": [
          "row_number",
          "code",
          "field",
          "message"
        ],
        "title": "ImportIssueOut",
        "type": "object"
      },
      "ImportOut": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "import_metadata": {
            "additionalProperties": true,
            "title": "Import Metadata",
            "type": "object"
          },
          "import_type": {
            "title": "Import Type",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "rows_accepted": {
            "title": "Rows Accepted",
            "type": "integer"
          },
          "rows_rejected": {
            "title": "Rows Rejected",
            "type": "integer"
          },
          "rows_total": {
            "title": "Rows Total",
            "type": "integer"
          },
          "source_status": {
            "$ref": "#/components/schemas/SourceStatus"
          },
          "status": {
            "$ref": "#/components/schemas/ImportStatus"
          }
        },
        "required": [
          "id",
          "provider",
          "import_type",
          "filename",
          "status",
          "source_status",
          "rows_total",
          "rows_accepted",
          "rows_rejected",
          "created_at",
          "completed_at",
          "error_message"
        ],
        "title": "ImportOut",
        "type": "object"
      },
      "ImportResultOut": {
        "properties": {
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "hit_events": {
            "title": "Hit Events",
            "type": "integer"
          },
          "is_duplicate": {
            "title": "Is Duplicate",
            "type": "boolean"
          },
          "metric_observations": {
            "title": "Metric Observations",
            "type": "integer"
          },
          "new_personal_records": {
            "title": "New Personal Records",
            "type": "integer"
          },
          "pitch_events": {
            "title": "Pitch Events",
            "type": "integer"
          },
          "raw_import_id": {
            "format": "uuid",
            "title": "Raw Import Id",
            "type": "string"
          },
          "rows_accepted": {
            "title": "Rows Accepted",
            "type": "integer"
          },
          "rows_rejected": {
            "title": "Rows Rejected",
            "type": "integer"
          },
          "rows_total": {
            "title": "Rows Total",
            "type": "integer"
          },
          "sessions_created": {
            "title": "Sessions Created",
            "type": "integer"
          },
          "sessions_updated": {
            "title": "Sessions Updated",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "sync_jobs_queued": {
            "title": "Sync Jobs Queued",
            "type": "integer"
          },
          "unresolved_players": {
            "items": {
              "type": "string"
            },
            "title": "Unresolved Players",
            "type": "array"
          }
        },
        "required": [
          "raw_import_id",
          "status",
          "is_duplicate",
          "sessions_created",
          "sessions_updated",
          "pitch_events",
          "hit_events",
          "metric_observations",
          "new_personal_records",
          "sync_jobs_queued",
          "rows_total",
          "rows_accepted",
          "rows_rejected",
          "unresolved_players"
        ],
        "title": "ImportResultOut",
        "type": "object"
      },
      "ImportStatus": {
        "description": "Lifecycle of a single ingestion attempt.",
        "enum": [
          "RECEIVED",
          "PROCESSING",
          "SUCCESS",
          "PARTIAL",
          "FAILED",
          "SKIPPED_DUPLICATE"
        ],
        "title": "ImportStatus",
        "type": "string"
      },
...
      "PlayerSummary": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "bats": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bats"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "graduation_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduation Year"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "throws": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Throws"
          }
        },
        "required": [
          "id",
          "display_name",
          "full_name",
          "first_name",
          "last_name",
          "active"
        ],
        "title": "PlayerSummary",
        "type": "object"
      },
...
      "ResolveIdentityIn": {
        "description": "Attach an unknown vendor athlete to one of our players.\n\n`player_id` is required and explicit: there is no \"match by name\" option,\nby design.",
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "player_id": {
            "format": "uuid",
            "title": "Player Id",
            "type": "string"
          },
          "reprocess": {
            "default": true,
            "title": "Reprocess",
            "type": "boolean"
          }
        },
        "required": [
          "player_id"
        ],
        "title": "ResolveIdentityIn",
        "type": "object"
      },
      "ResolveIdentityOut": {
        "description": "What a mapping actually recovered.",
        "properties": {
          "imports_reprocessed": {
            "title": "Imports Reprocessed",
            "type": "integer"
          },
          "new_personal_records": {
            "title": "New Personal Records",
            "type": "integer"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerSummary"
          }
        },
        "required": [
          "player",
          "imports_reprocessed",
          "new_personal_records"
        ],
        "title": "ResolveIdentityOut",
        "type": "object"
      },
...
      "SourceStatus": {
        "description": "Confidence level of vendor-supplied data.\n\nTrackMan can publish a session as PRELIMINARY and later republish it as\nVERIFIED with corrected values. Everything derived from an event carries the\nstatus of the data it was derived from so the UI can label it honestly.",
        "enum": [
          "PRELIMINARY",
          "VERIFIED"
        ],
        "title": "SourceStatus",
        "type": "string"
      },
...
      "UnresolvedIdentityOut": {
        "properties": {
          "external_display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Display Name"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "occurrence_count": {
            "title": "Occurrence Count",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider",
          "external_id",
          "external_display_name",
          "occurrence_count",
          "last_seen_at"
        ],
        "title": "UnresolvedIdentityOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Internal player analytics for a baseball training facility. All athlete access is authorized server-side and scoped to one organization.",
    "title": "Baseball Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/identity/unresolved": {
      "get": {
        "description": "Vendor athletes awaiting a human decision.",
        "operationId": "list_unresolved_api_v1_identity_unresolved_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UnresolvedIdentityOut"
                  },
                  "title": "Response List Unresolved Api V1 Identity Unresolved Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Unresolved",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/identity/unresolved/{item_id}/resolve": {
      "post": {
        "description": "Map a vendor athlete to one of our players, and recover their held data.\n\nAdmin-only and audited: this decision determines whose data is whose, and it\nis the one place a mistake would attribute one athlete's numbers to another.\n\nEvents held back for this athlete are recovered by **reprocessing** the\naffected imports from their archived source bytes -- never by back-filling\nrows by hand. The recovered data therefore goes through exactly the same\nvalidated pipeline as a first ingest, and stays idempotent.",
        "operationId": "resolve_identity_api_v1_identity_unresolved__item_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveIdentityIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveIdentityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Identity",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports": {
      "get": {
        "operationId": "list_imports_api_v1_imports_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImportOut"
                  },
                  "title": "Response List Imports Api V1 Imports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Imports",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/trackman/csv": {
      "post": {
        "description": "Ingest a TrackMan CSV export.\n\nRe-uploading the same file is safe and explicitly reported as a duplicate\nrather than silently creating a second copy of the session.",
        "operationId": "upload_trackman_csv_api_v1_imports_trackman_csv_post",
        "parameters": [
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_trackman_csv_api_v1_imports_trackman_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Trackman Csv",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}": {
      "get": {
        "description": "One import with its rejected rows, so nothing is silently discarded.",
        "operationId": "get_import_api_v1_imports__import_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Import",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{import_id}/reprocess": {
      "post": {
        "description": "Re-run a stored import from its archived source bytes.\n\nUsed after mapping a previously-unknown athlete, and after a parser or\nmetric change. Audited, because it rewrites athlete-facing numbers.",
        "operationId": "reprocess_import_api_v1_imports__import_id__reprocess_post",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reprocess Import",
        "tags": [
          "imports"
        ]
      }
    },
...
  }
}